Add the MVP to the list of "finished proposals". (#210)
The MVP wasn't a typical "proposal", however it was the source of many
features in the specs today, in the same way that subsequent feature
proposals are. And like all the other feature proposals, the MVP was
designed in the CG, approved by the WG, and it had affected specs.

This makes the table complete in a sense, as it documents all the
moments when features were added to specs, including the web-api spec
which wasn't previously present in the table.